+++ pyweber/models/__init__.py
@@ -1,6 +1,6 @@
-from .request import Headers, Request, File, request
+from .request import Headers, Request, File, FieldStorage, Field, request
 from .response import Response
 from .run import run, run_as_asgi
 from .routes import Route, RedirectRoute
 
-__all__ = ['Headers', 'Request', 'request', 'File', 'Response', 'RedirectRoute', 'Route', 'run', 'run_as_asgi']
+__all__ = ['Headers', 'Request', 'request', 'File', 'FieldStorage', 'Field', 'Response', 'RedirectRoute', 'Route', 'run', 'run_as_asgi']