+++ pyweber/models/cookies.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import datetime, timezone, timedelta
 
 class CookieManager:
     def __init__(self):
@@ -16,7 +16,10 @@ class CookieManager:
         samesite: str = 'Strict',
         httponly: bool = True,
         secure: bool = True,
-        expires: datetime = None
+        expires_after_seconds: int = None,
+        expires_after_hours: int = None,
+        expires_after_days: int = None,
+        max_age: int = None
     ):
         cookie = f'{cookie_name}={cookie_value}; Path={path};'
 
@@ -26,17 +29,34 @@ class CookieManager:
         if secure:
             cookie += ' Secure;'
         
-        if samesite not in ['Strict', 'Lax', None]:
-            raise AttributeError('Samsite is not valid. Please use one of the options: [Strict, Lax, None]')
-        
-        if expires and not isinstance(expires, datetime):
-            raise ArithmeticError('Datetime is not valid, please use datetime to define the expires date.')
+        if samesite:
+            if samesite not in ['Strict', 'Lax']:
+                raise ValueError("SameSite is not valid. Please use one of: ['Strict', 'Lax']")
 
-        if expires:
+            cookie += f' SameSite={str(samesite)}'
+
+        expires_after_days = expires_after_days if isinstance(
+            expires_after_days, (int, float)
+        ) and expires_after_days > 0 else 0
+        expires_after_hours = expires_after_hours if isinstance(
+            expires_after_hours, (int, float)
+        ) and expires_after_hours > 0 else 0
+        expires_after_seconds = expires_after_seconds if isinstance(
+            expires_after_seconds, (int, float)
+        ) and expires_after_seconds > 0 else 0
+
+        if expires_after_days>0 or expires_after_hours>0 or expires_after_seconds>0:
+            expires = datetime.now(timezone.utc) + timedelta(
+                days=expires_after_days,
+                seconds=expires_after_seconds,
+                hours=expires_after_hours
+            )
+
             expires_str = expires.strftime("%a, %d %b %Y %H:%M:%S GMT")
             cookie += f' Expires={expires_str};'
-    
-        cookie += f' SameSite={str(samesite)}'
+        
+        if isinstance(max_age, (int, float)) and max_age > 0:
+            cookie += f' Max-Age={max_age};'
         
         if cookie not in self.__cookies:
             self.__cookies.append(cookie)